requirements.txt
Django==4.1.5
git+https://github.com/mercadolibre/python-sdk.git
python-decouple==3.6
